requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
clip>=0.15.0
easyocr
transformers
fastapi
uvicorn
python-multipart
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0